Todo: GitHub Actions

+++ nogizaka.py
@@ -16,12 +16,19 @@ def add_host(str: str) -> str:
 
 def download_image_return_path(img_src_url: str, repo_name: str) -> str:
     img_full_url = add_host(img_src_url)
-    response = requests.get(img_full_url)
     img_relative_path = f"{repo_name}/{urllib.parse.urlparse(img_full_url).path[1:]}"
+    if os.path.exists(img_relative_path):
+        print(f"File exists: {img_relative_path}")
+        return "/" + img_relative_path
     os.makedirs(os.path.dirname(img_relative_path), exist_ok=True)
-    with open(img_relative_path, "wb") as f:
-        f.write(response.content)
-    return "/" + img_relative_path
+    while True:
+        try:
+            response = requests.get(img_full_url)
+            with open(img_relative_path, "wb") as f:
+                f.write(response.content)
+            return "/" + img_relative_path
+        except:
+            pass
 
 
 def get_profile(member_id: int):
@@ -44,10 +51,10 @@ def get_profile(member_id: int):
     dl_list = soup.find_all("dl", class_="md--hd__data a--tx js-tdi")
     for dl in dl_list:
         key = dl.find_all("dt")[0].get_text()
+        result["SNS"] = {}
         if key != "SNS":
             result[key] = dl.find_all("dd")[0].get_text()
         else:
-            result["SNS"] = {}
             for a in dl.find_all("dd")[0].children:
                 result["SNS"][a.get_text()] = a.get("href")
 
@@ -72,7 +79,14 @@ def get_articles_url_list(member_id: int):
         # print(soup.prettify())
         a_list = soup.find_all("a", class_="bl--card js-pos a--op hv--thumb")
         for a in a_list:
-            articles_url_list.append(add_host(a.get("href")))
+            url_with_param = add_host(a.get("href"))
+            # Parse the URL
+            parsed_url = urllib.parse.urlparse(url_with_param)
+            # Reconstruct the URL without the query parameters
+            url_without_params = urllib.parse.urlunparse(
+                (parsed_url.scheme, parsed_url.netloc, parsed_url.path, "", "", "")
+            )
+            articles_url_list.append(url_without_params)
         print(f"{len(a_list)} results on page {current_page}.")
 
         next_li = soup.find_all("li", class_="next")
@@ -93,27 +107,33 @@ def get_articles_url_list(member_id: int):
 
 
 def get_blog_content(url: str, repo_name: str):
-    data = {}
-    soup = BeautifulSoup(requests.get(url).content, "lxml")
-    data["title"] = soup.find_all("h1", class_="bd--hd__ttl f--head a--tx js-tdi")[
-        0
-    ].get_text()
+    while True:
+        try:
+            data = {}
+            soup = BeautifulSoup(requests.get(url).content, "lxml")
+            data["title"] = soup.find_all(
+                "h1", class_="bd--hd__ttl f--head a--tx js-tdi"
+            )[0].get_text()
 
-    data["time"] = soup.find_all("p", class_="bd--hd__date a--tx js-tdi")[0].get_text()
+            data["time"] = soup.find_all("p", class_="bd--hd__date a--tx js-tdi")[
+                0
+            ].get_text()
 
-    data["url"] = url
+            data["url"] = url
 
-    content = soup.find_all("div", class_="bd--edit")[0]
-    img_list = content.find_all("img")
-    for img in img_list:
-        # print(img)
-        # # https://www.nogizaka46.com/s/n46/diary/detail/100646 empty img no src
-        if img.get("src"):
-            img["src"] = download_image_return_path(img.get("src"), repo_name)
+            content = soup.find_all("div", class_="bd--edit")[0]
+            img_list = content.find_all("img")
+            for img in img_list:
+                # print(img)
+                # # https://www.nogizaka46.com/s/n46/diary/detail/100646 empty img no src
+                if img.get("src"):
+                    img["src"] = download_image_return_path(img.get("src"), repo_name)
 
-    data["content"] = str(content)
+            data["content"] = str(content)
 
-    return data
+            return data
+        except:
+            pass
 
 
 def main(member_id: int):
@@ -129,9 +149,10 @@ def main(member_id: int):
         print(f"downloading {i+1}/{len(articles_url_list)}: {articles_url_list[i]}")
         article = get_blog_content(articles_url_list[i], result["repo_name"])
         result["blog"].append(article)
-        time.sleep(random.randint(1, 3))
+        # time.sleep(random.randint(1, 3))
     with open(f"{result['repo_name']}/result.json", "w", encoding="utf-8") as f:
         json.dump(result, f, ensure_ascii=False)
 
 
-main(48012)  # Hayakawa Seira
+# main(36754)
+main(48009)
